% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_series.R
\name{add_bars}
\alias{add_bars}
\title{Add Bar Series Data to the SveltePlots Chart}
\usage{
add_bars(sp, current_data, color_mapping, group_name, include_legend)
}
\arguments{
\item{sp}{A SveltePlot htmlwidget object to which the series will be added.}

\item{current_data}{The current data frame which contains the series.}

\item{color_mapping}{A named list mapping group names to colors.}

\item{group_name}{The names of the series.}

\item{include_legend}{A logical value indicating whether to include the data series in the legend.}
}
\value{
Configuration list with modified data for the bar chart.
}
\description{
This function adds series to a SveltePlot chart, updating the internal data structure.
It supports grouped and stacked bar charts with customization of colors and legend entries.
}
\keyword{internal}
