% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_helpers.R
\name{bar_chart}
\alias{bar_chart}
\title{Internal: Prepare Bar Chart Data}
\usage{
bar_chart(
  data,
  create_configurations,
  size,
  alpha,
  tooltip,
  colors,
  mode,
  include_legend,
  facet_var
)
}
\arguments{
\item{data}{Data frame to be processed.}

\item{create_configurations}{Function to create default configurations for charts.}

\item{size}{alpha, tooltip Visual properties for the chart.}

\item{colors}{Custom colors for the chart, if specified.}

\item{mode}{Bar chart mode ("stacked", "grouped", "percent").}

\item{include_legend}{Logical indicating whether to include a legend.}
}
\value{
Configuration list with modified data for bar chart.
}
\description{
Handles data preparation for bar charts with modes such as stacked or grouped.
}
\keyword{internal}
