% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CO2}
\alias{CO2}
\title{Plant Growth Dataset}
\format{
A tibble with 84 rows and 6 variables:
\describe{
  \item{Plant}{Factor. The identifier for the plant.}
  \item{Type}{Factor. The type of plant, here represented as "Quebec".}
  \item{Treatment}{Factor. The treatment applied to the plant, here represented as "nonchilled".}
  \item{conc}{Numeric. The concentration level of the treatment.}
  \item{uptake}{Numeric. The uptake measurement of the plant.}
  \item{date}{Date. The date of the observation.}
}
}
\usage{
CO2
}
\description{
This dataset contains information about the growth of plants under different treatments and concentrations over time.
}
\examples{
data(CO2)
head(CO2)
}
\keyword{datasets}
