% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{confidence_intervals}
\alias{confidence_intervals}
\title{Confidence Intervals Data}
\format{
A tibble with 368 rows and 7 variables:
\describe{
  \item{.model_id}{Integer. The model ID.}
  \item{.model_desc}{Character. Description of the model.}
  \item{.key}{Factor. The key indicating actual or forecast data.}
  \item{.index}{Date. The date of the observation.}
  \item{.value}{Numeric. The actual value.}
  \item{.conf_lo}{Numeric. The lower bound of the confidence interval.}
  \item{.conf_hi}{Numeric. The upper bound of the confidence interval.}
}
}
\usage{
confidence_intervals
}
\description{
This dataset contains time series data with actual values and corresponding confidence intervals.
}
\examples{
data(confidence_intervals)
head(confidence_intervals)
}
\keyword{datasets}
