% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dau}
\alias{dau}
\title{Daily Active Users Data}
\format{
A tibble with 19 rows and 3 variables:
\describe{
  \item{date}{The date, in Date format.}
  \item{DAU}{Daily active users count, a numeric value.}
  \item{Slot}{Slot number associated with the DAU count, a numeric value.}
}
}
\usage{
dau
}
\description{
This dataset contains daily active user (DAU) counts for a specific period, along with slot information.
}
\examples{
data(dau)
head(dau)
}
\keyword{datasets}
