% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_helpers.R
\name{density_chart}
\alias{density_chart}
\title{Internal: Prepare Density Chart Data}
\usage{
density_chart(
  data,
  create_configurations,
  size,
  alpha,
  tooltip,
  colors,
  include_legend,
  facet_var
)
}
\arguments{
\item{data}{Data frame to be processed.}

\item{create_configurations}{Function to create default configurations for charts.}

\item{size}{alpha, tooltip Visual properties for the chart.}

\item{colors}{Custom colors for the chart, if specified.}

\item{include_legend}{Logical indicating whether to include a legend.}
}
\value{
Configuration list with modified data for density chart.
}
\description{
Transforms data for rendering density charts. This function is part of the internal mechanics
for data preparation in SveltePlot charts.
}
\keyword{internal}
