% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economics}
\alias{economics}
\title{Economics Data}
\format{
A tibble with 574 rows and 6 variables:
\describe{
  \item{date}{Date. The date of the observation.}
  \item{pce}{Numeric. Personal consumption expenditures, in billions of dollars.}
  \item{pop}{Numeric. Total population, in thousands.}
  \item{psavert}{Numeric. Personal savings rate, as a percentage.}
  \item{uempmed}{Numeric. Median duration of unemployment, in weeks.}
  \item{unemploy}{Numeric. Number of unemployed individuals, in thousands.}
}
}
\source{
This dataset is sourced from the Federal Reserve Economic Data (FRED) database.
\url{https://fred.stlouisfed.org/}
}
\usage{
economics
}
\description{
This dataset contains economic data, including personal consumption expenditures, population, personal savings rate, median duration of unemployment, and the number of unemployed individuals, recorded monthly from July 1967.
}
\examples{
data(economics)
head(economics)
}
\keyword{datasets}
