% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fruit}
\alias{fruit}
\title{Example Dataset: Fruit Data}
\format{
A tibble with 16 rows and 6 variables:
\describe{
  \item{id}{Integer. A unique identifier for each record.}
  \item{year}{Character. The year the data was recorded.}
  \item{basket}{Integer. The basket number containing the fruits.}
  \item{fruit}{Character. The type of fruit.}
  \item{value}{Integer. The value associated with the fruit.}
  \item{color}{Character. The color of the fruit.}
}
}
\usage{
fruit
}
\description{
This dataset contains information about different types of fruits in various baskets over multiple years. It includes details such as the fruit type, the value, and the color of the fruit.
}
\examples{
data(fruit)
head(fruit)

}
\keyword{datasets}
