% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{Gapminder Dataset}
\format{
A data frame with 1704 rows and 6 variables.
\describe{
  \item{country}{Factor with country names.}
  \item{continent}{Factor with continent names.}
  \item{year}{Integer.}
  \item{lifeExp}{Numeric.}
  \item{pop}{Integer.}
  \item{gdpPercap}{Numeric.}
}
}
\source{
This dataset is sourced from the `gapminder` package.
}
\usage{
gapminder
}
\description{
This dataset provides country-level data on life expectancy, GDP per capita, and population. It is included in the `gapminder` package.
For detailed information, please refer to the [gapminder documentation](https://cran.r-project.org/package=gapminder).
}
\examples{
data(gapminder)
head(gapminder)
}
\seealso{
\href{https://cran.r-project.org/package=gapminder}{gapminder documentation}
}
\keyword{datasets}
