% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_helpers.R
\name{histogram}
\alias{histogram}
\title{Internal: Prepare Histogram Data}
\usage{
histogram(
  data,
  create_configurations,
  size,
  alpha,
  tooltip,
  colors,
  breaks,
  include_legend,
  facet_var
)
}
\arguments{
\item{data}{Data frame to be processed.}

\item{create_configurations}{Function to create default configurations for charts.}

\item{size}{alpha, tooltip Visual properties for the chart.}

\item{colors}{Custom colors for the chart, if specified.}

\item{breaks}{Break points for histograms.}

\item{include_legend}{Logical indicating whether to include a legend.}
}
\value{
Configuration list with modified data for histogram.
}
\description{
Transforms and bins data for histogram chart rendering within the SveltePlot package.
}
\keyword{internal}
