% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_helpers.R
\name{mutate_mapping}
\alias{mutate_mapping}
\title{Internal: Mutate Mapping for Chart Data}
\usage{
mutate_mapping(
  data,
  mapping,
  create_configurations,
  size,
  alpha,
  tooltip,
  colors = NULL,
  mode,
  include_legend,
  breaks,
  facet_var
)
}
\arguments{
\item{data}{Data frame to be processed.}

\item{mapping}{Aesthetic mappings created by `spaes()`.}

\item{create_configurations}{Function to create default configurations for charts.}

\item{size, }{alpha, tooltip Visual properties for the chart.}

\item{colors}{Custom colors for the chart, if specified.}

\item{mode}{Specific mode for bar charts ("stacked", "grouped", etc.).}

\item{include_legend}{Logical indicating whether to include a legend.}

\item{breaks}{Break points for histograms.}
}
\value{
Modified chart configurations.
}
\description{
Prepares and modifies data mappings for SveltePlot chart rendering. This function is internally used
by the `sp()` function to handle data transformation based on chart type and aesthetic mappings.
}
\keyword{internal}
