% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Penguins Dataset}
\format{
A data frame with 344 rows and 8 variables.
\describe{
  \item{species}{Factor with levels Adelie, Chinstrap, Gentoo.}
  \item{island}{Factor with levels Biscoe, Dream, Torgersen.}
  \item{bill_length_mm}{Numeric.}
  \item{bill_depth_mm}{Numeric.}
  \item{flipper_length_mm}{Numeric.}
  \item{body_mass_g}{Numeric.}
  \item{sex}{Factor with levels female, male.}
  \item{year}{Integer.}
}
}
\source{
This dataset is sourced from the `palmerpenguins` package.
}
\usage{
penguins
}
\description{
This dataset provides measurements of penguins. It is included in the `palmerpenguins` package.
For detailed information, please refer to the [palmerpenguins documentation](https://allisonhorst.github.io/palmerpenguins/).
}
\examples{
data(penguins)
head(penguins)
}
\seealso{
\href{https://allisonhorst.github.io/palmerpenguins/}{palmerpenguins documentation}
}
\keyword{datasets}
