% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{purchases}
\alias{purchases}
\title{Revenue Data}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{date}{Date. The timestamp of the revenue data.}
  \item{age}{Factor. The age range category.}
  \item{revenue}{Numeric. The revenue for the given date.}
  \item{revenue_roll}{Numeric. The rolling revenue calculation.}
}
}
\usage{
purchases
}
\description{
This dataset contains revenue data over a period of time with rolling revenue calculations.
}
\examples{
data(purchases)
head(purchases)
}
\keyword{datasets}
