% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{purchases_age}
\alias{purchases_age}
\title{Revenue Data}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{adjusted_timestamp}{Date. The timestamp of the revenue data.}
  \item{age}{Factor. The age range category.}
  \item{revenue}{Numeric. The revenue for the given date.}
  \item{revenue_roll}{Numeric. The rolling revenue calculation.}
}
}
\usage{
purchases_age
}
\description{
This dataset contains revenue data over a period of time with rolling revenue calculations.
}
\examples{
data(purchases_age)
head(purchases_age)
}
\keyword{datasets}
