% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quests}
\alias{quests}
\title{Example Dataset: Retention Data}
\format{
A tibble with 38 rows and 5 variables:
\describe{
  \item{custom_01}{Character. A custom category identifier.}
  \item{progression_2}{Factor. The progression stage.}
  \item{n.x}{Numeric. The number of observations at the first point.}
  \item{n.y}{Numeric. The number of observations at the second point.}
  \item{retention}{Numeric. The retention rate, calculated as n.x divided by n.y.}
}
}
\usage{
quests
}
\description{
This dataset contains retention data for different custom categories and progression stages. It includes the number of observations at two different points (n.x and n.y) and the retention rate.
}
\examples{
data(quests)
head(quests)

}
\keyword{datasets}
