% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-api.R
\name{sp_title}
\alias{sp_title}
\title{Set title for SveltePlots charts}
\usage{
sp_title(
  sp,
  title = NULL,
  text_align = "left",
  color = "black",
  font_size = 16,
  font_weight = "bold",
  custom_css = ""
)
}
\arguments{
\item{sp}{The SveltePlots object to modify.}

\item{title}{The title text to be displayed.}

\item{text_align}{Text alignment of the title ("left", "center", "right"; default: "left").}

\item{color}{Color of the title text (default: "black").}

\item{font_size}{Font size of the title text (default: 16).}

\item{font_weight}{Font weight of the title text (default: "bold").}

\item{custom_css}{Some custom css for the title.}
}
\description{
This function sets the title properties for SveltePlots charts, including text alignment, color, font size, font weight, and padding.
}
