% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SveltePlots.R
\name{SveltePlots-shiny}
\alias{SveltePlots-shiny}
\alias{SveltePlotsOutput}
\alias{renderSveltePlots}
\title{Shiny bindings for SveltePlots}
\usage{
SveltePlotsOutput(outputId, width = "100\%", height = "400px")

renderSveltePlots(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a AgeGroupFacet}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
No return value. Called for side effects in a 'shiny' app context.

No return value. Called for side effects in a 'shiny' app context.
}
\description{
Output and render functions for using SveltePlots within Shiny
applications and interactive Rmd documents.
}
