% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SveltePlots.R
\name{SveltePlots}
\alias{SveltePlots}
\title{Chart Specification}
\usage{
SveltePlots(list_input)
}
\arguments{
\item{list_input}{This is a list inside a list or a list of lists which contain the chart specifications used by Svelte to generate the chart.}
}
\description{
Chart Specification
}
\keyword{internal}
